chore: Change folder structure and naming
Applied standard naming convention to all files and folders. Also arranged the structure of the folders.

feat: Add responsive feature to the application

This commit applies a responsive feature to the entire application and significant changes in the view at screens below 600px.

feat: Add responsive navbar

This commit adds a collapsible navbar with a hamburger icon.